+++ app/src/main/java/ru/easycode/words504/admintools/presentation/WordsIndexesInputLayout.kt
@@ -1,14 +1,14 @@
-package ru.easycode.words504.admintools
+package ru.easycode.words504.admintools.presentation
 
 import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import com.google.android.material.textfield.TextInputLayout
 import ru.easycode.words504.R
-import ru.easycode.words504.admintools.input.ApostropheMatch
-import ru.easycode.words504.admintools.input.HandleIndexes
-import ru.easycode.words504.admintools.input.LetterMatch
-import ru.easycode.words504.admintools.input.WordsIndexesTextWatcher
+import ru.easycode.words504.admintools.presentation.input.ApostropheMatch
+import ru.easycode.words504.admintools.presentation.input.HandleIndexes
+import ru.easycode.words504.admintools.presentation.input.LetterMatch
+import ru.easycode.words504.admintools.presentation.input.WordsIndexesTextWatcher
 import ru.easycode.words504.domain.Mapper
 
 class WordsIndexesInputLayout : TextInputLayout, Mapper.Unit<String> {
