+++ app/src/main/java/ru/easycode/words504/admintools/presentation/input/EditTextWatcher.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.admintools.input
+package ru.easycode.words504.admintools.presentation.input
 
 import android.text.Editable
 import android.text.TextWatcher
